% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{c.chars}
\alias{c.chars}
\title{Combine \code{chars} Objects}
\usage{
\method{c}{chars}(...)
}
\arguments{
\item{...}{\code{chars} objects.}
}
\value{
a larger \code{chars} object containing the combined elements of \code{...}.
}
\description{
Combine \code{chars} Objects
}
\examples{
c(chars("java"), chars("script"))
}
