% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charcuterie-package.R
\docType{package}
\name{charcuterie-package}
\alias{charcuterie}
\alias{charcuterie-package}
\title{charcuterie: Handle Strings as Vectors of Characters}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Creates a new chars class which looks like a string but is actually a vector of single characters.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jonocarroll/charcuterie}
  \item Report bugs at \url{https://github.com/jonocarroll/charcuterie/issues}
}

}
\author{
\strong{Maintainer}: Jonathan Carroll \email{rpkg@jcarroll.com.au} (\href{https://orcid.org/0000-0002-1404-5264}{ORCID})

}
\keyword{internal}
