% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chars.R
\name{chars.default}
\alias{chars.default}
\title{Convert an Object to \code{chars}}
\usage{
\method{chars}{default}(x, ...)
}
\arguments{
\item{x}{object to convert}

\item{...}{other options}
}
\value{
\code{.NotYetImplemented()} error
}
\description{
Convert an Object to \code{chars}
}
