% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chars.R
\name{chars}
\alias{chars}
\title{Create a \code{chars} Object}
\usage{
chars(x, ...)
}
\arguments{
\item{x}{object to convert to \code{chars}.}

\item{...}{other options passed to methods.}
}
\value{
an object of class \code{chars}.
}
\description{
Create a \code{chars} Object
}
