% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{count}
\alias{count}
\title{Count Characters in a Chars Object}
\usage{
count(x, value, ignore.case = FALSE)
}
\arguments{
\item{x}{A vector of characters.}

\item{value}{character (length 1) to count}

\item{ignore.case}{should the letter case be ignored?}
}
\value{
integer, count of instances of \code{value} in \code{x}
}
\description{
Count Characters in a Chars Object
}
\examples{
count(chars("strawberry"), 3)
}
