% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/except.R
\name{except}
\alias{except}
\title{Elements of x except those in y}
\usage{
except(x, y)
}
\arguments{
\item{x}{larger vector}

\item{y}{smaller vector}
}
\description{
Does not treat the operation as a set
}
\examples{
except(c(1:5), 3)
except(chars("abcde"), "c")
except(chars("abracadabra"), "b")
}
