% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{format.chars}
\alias{format.chars}
\title{Format a \code{chars} Object}
\usage{
\method{format}{chars}(x, ...)
}
\arguments{
\item{x}{a \code{chars} object.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
a formatted \code{chars} object.
}
\description{
Format a \code{chars} Object
}
