% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{head.chars}
\alias{head.chars}
\title{Return the First Parts of a \code{chars} Object}
\usage{
\method{head}{chars}(x, ...)
}
\arguments{
\item{x}{a \code{chars} object.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
the first (\code{n}) elements of a \code{chars} object as a \code{chars} object.
}
\description{
Return the First Parts of a \code{chars} Object
}
\examples{
head(chars("abcdefhgi"))

head(chars("javascript"), 4)
}
