% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{intersect.chars}
\alias{intersect.chars}
\title{Setwise Intersection of \code{chars} Objects}
\usage{
\method{intersect}{chars}(x, y, ...)
}
\arguments{
\item{x}{a \code{chars} object.}

\item{y}{a \code{chars} object or character vector.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
the setwise intersection of \code{x} and \code{y}.
}
\description{
Setwise Intersection of \code{chars} Objects
}
\examples{
union(chars("pine"), chars("apple"))
}
