% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{is_alnum}
\alias{is_alnum}
\title{Is a Character a Letter OR a Number?}
\usage{
is_alnum(x)
}
\arguments{
\item{x}{A vector of characters.}
}
\value{
A boolean vector indicating whether each element of \code{x} is a
letter or a number.
}
\description{
A combination of \code{\link[=is_letter]{is_letter()}} and \code{\link[=is_number]{is_number()}}.
}
\examples{
is_alnum(chars("Lee7c0deR 4 L1fe"))

Filter(is_alnum, chars("2 B or !2 B"))
}
