% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{is_letter}
\alias{is_letter}
\title{Is a Character a Letter?}
\usage{
is_letter(x)
}
\arguments{
\item{x}{A vector of characters.}
}
\value{
A boolean vector indicating whether each element of \code{x} is a
letter (appears in \code{letters} ignoring case).
}
\description{
Compares against the values of \code{letters} (the English alphabet),
ignoring case.
}
\examples{
is_letter(chars("Lee7c0deR"))

Filter(is_letter, chars("w00t"))
}
