% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{is_number}
\alias{is_number}
\title{Is a Character a Number?}
\usage{
is_number(x)
}
\arguments{
\item{x}{A vector of characters.}
}
\value{
A boolean vector indicating whether each element of \code{x} is a
number (appears in \code{0:9} as a number)
}
\description{
Compares against the values of \code{0:9} (as a number).
}
\examples{
is_number(chars("Lee7c0deR"))

Filter(is_number, chars("w00t"))
}
