% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{is_punct}
\alias{is_punct}
\title{Is a Character Punctuation?}
\usage{
is_punct(x)
}
\arguments{
\item{x}{A vector of characters.}
}
\value{
A boolean vector indicating whether each element of \code{x} is
considered as punctuation.
}
\description{
Compares against the regex group \verb{[[:punct:]]}.
}
\examples{
is_punct(chars("I can haz?"))

Filter(Negate(is_punct), chars("abc,123;$*\%?"))
}
