% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{setdiff}
\alias{setdiff}
\alias{union}
\alias{intersect}
\title{Set Operations}
\usage{
setdiff(x, y)

union(x, y)

intersect(x, y)
}
\arguments{
\item{x}{vector}

\item{y}{vector}
}
\description{
Set Operations

Set Operations

Set Operations
}
