% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.chars}
\alias{print.chars}
\title{Print a \code{chars} Object}
\usage{
\method{print}{chars}(x, ...)
}
\arguments{
\item{x}{a \code{chars} object.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
\code{x} (invisibly), used to print to console.
}
\description{
Print a \code{chars} Object
}
