% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{rev.chars}
\alias{rev.chars}
\title{Reverse Elements of a \code{chars} Object}
\usage{
\method{rev}{chars}(x)
}
\arguments{
\item{x}{a \code{chars} object}
}
\value{
a \code{chars} object with the elements reversed.
}
\description{
Reverse Elements of a \code{chars} Object
}
\examples{
rev(chars("racecar"))

rev(chars("alphabet"))
}
