% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{setdiff.chars}
\alias{setdiff.chars}
\title{Setwise Difference Between \code{chars} Objects}
\usage{
\method{setdiff}{chars}(x, y, ...)
}
\arguments{
\item{x}{a \code{chars} object.}

\item{y}{a \code{chars} object or character vector.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
the setwise difference of \code{x} and \code{y}.
}
\description{
Setwise Difference Between \code{chars} Objects
}
\examples{
setdiff(chars("javascript"), chars("script"))
}
