% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{sort.chars}
\alias{sort.chars}
\title{Sort a \code{chars} Object}
\usage{
\method{sort}{chars}(x, decreasing = FALSE, ...)
}
\arguments{
\item{x}{a \code{chars} object.}

\item{decreasing}{logical. Should the sort be increasing or decreasing? Not available for partial sorting.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
a sorted \code{chars} object.
}
\description{
Sort a \code{chars} Object
}
\examples{
sort(chars("alphabet"))
}
