% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chars.R
\name{string}
\alias{string}
\title{Create a String From a \code{chars} Object}
\usage{
string(x, collapse = "", ...)
}
\arguments{
\item{x}{one or more \code{chars} objects.}

\item{collapse}{an optional character string to separate the results. Not NA_character_.}

\item{...}{other arguments passed to \code{\link[=paste]{paste()}}}
}
\value{
a \code{character} (traditional R string) with the elements of \code{x} in a single value.
}
\description{
Create a String From a \code{chars} Object
}
