% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{[.chars}
\alias{[.chars}
\title{Extract or Replace Parts of a \code{chars} Object}
\usage{
\method{[}{chars}(x, ...)
}
\arguments{
\item{x}{a \code{chars} object.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
the extracted parts of a \code{chars} object, or a \code{chars} object with
replacements performed.
}
\description{
Extract or Replace Parts of a \code{chars} Object
}
\examples{
s <- chars("censor")
s[2:5]
s[2:5] <- "X"
s
}
