% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{tail.chars}
\alias{tail.chars}
\title{Return the Last Parts of a \code{chars} Object}
\usage{
\method{tail}{chars}(x, ...)
}
\arguments{
\item{x}{a \code{chars} object.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
the last (\code{n}) elements of a \code{chars} object as a \code{chars} object.
}
\description{
Return the Last Parts of a \code{chars} Object
}
\examples{
tail(chars("javascript"))

tail(chars("abcdefghi"))
}
