% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{union.chars}
\alias{union.chars}
\title{Setwise Union of \code{chars} Objects}
\usage{
\method{union}{chars}(x, y, ...)
}
\arguments{
\item{x}{a \code{chars} object.}

\item{y}{a \code{chars} object or character vector.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
the setwise union of \code{x} and \code{y}.
}
\description{
Setwise Union of \code{chars} Objects
}
\examples{
union(chars("java"), chars("script"))
}
