% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{unique.chars}
\alias{unique.chars}
\title{Extract Unique Elements of \code{chars} Objects.}
\usage{
\method{unique}{chars}(x, ...)
}
\arguments{
\item{x}{a \code{chars} object.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
a \code{chars} object containing unique elements.
}
\description{
Extract Unique Elements of \code{chars} Objects.
}
\examples{
unique(chars("mississippi"))
}
