% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chars.R
\name{chars.character}
\alias{chars.character}
\title{Create a chars object from a string}
\usage{
\method{chars}{character}(x, ...)
}
\arguments{
\item{x}{string to convert to a chars object (length 1 only)}

\item{...}{unused}
}
\description{
Create a chars object from a string
}
\details{
\code{chars} expects a single string as input. To create a list of these,
consider \code{lapply(strings, chars)}
}
