% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chars.R
\name{string}
\alias{string}
\title{Create a string from a chars object}
\usage{
string(..., collapse = "")
}
\arguments{
\item{...}{one or more \R objects, to be converted to character vectors.}

\item{collapse}{an optional character string to separate the results.  Not
    \code{\link[base]{NA_character_}}.  When \code{collapse} is a string,
    the result is always a string (\code{\link[base]{character}} of length 1).}
}
\description{
Create a string from a chars object
}
